modify layout

+++ pages/cards/_id.vue
@@ -2,20 +2,21 @@
   <div>
     <Header />
 
-    <div class="container">
+    <div class="container position-relative">
+      <div class="loading"></div>
       <div class="row justify-content-center">
         <div class="col-12 col-sm-8">
           <b-card
             :img-src="card.image"
             img-top
             tag="card"
-            class="text-center shadow-sm"
+            class="text-center"
           >
-            <div class="float-right small">
-              <nuxt-link :to="{ name: 'cards-edit-id', params: { id: card.id, card: card } }">編集</nuxt-link> /
-              <b-form @submit.prevent="deleteCard" class="d-inline">
-                <b-button type="submit" variant="link">削除</b-button>
-              </b-form>
+            <div class="position-relative">
+              <div class="position-top-right">
+                <nuxt-link :to="{ name: 'cards-edit-id', params: { id: card.id, card: card } }">編集</nuxt-link> /
+                <a href="#" @click.prevent="deleteCard()">削除</a>
+              </div>
             </div>
             <b-card-title>{{ card.title }}</b-card-title>
             <b-card-text>
@@ -48,6 +49,10 @@ export default {
   },
   methods: {
     deleteCard() {
+      if (!confirm('本当に削除しますか？')) return
+
+      document.querySelector('.loading').style.display = 'block';
+
       const target = this.$route.params.card.id
       this.$store.dispatch('deleteCard', { target })
         .then(() => {
